Update single_responsibility.md
Looks really great!  I do agree with Kyle that the Metz quote is your best hook.  What would it look like to lead with that?
At the end your could return to it.  Show how there are no "ands" in your description of the classes and how the two classes would change for two reasons.  I think that will book-end the piece nicely.